Add search-using-default
search-using-default = (items, callback, default-callback) ->

This is the same as the normal search function, except if there are no
matches the default callback is called with the current query string
("needle" in the source) as an argument.

If default-callback is not provided, the normal callback is used.

Going forward it might be nice to also provide an option for an exit
callback rather than killing the process. -POLM